test: adds tests to requestBodyAsString Utility method (#13)
Also includes refactoring of code, moving general functions out of the server class.
