+++ src/main/java/group8/server/ContinuousIntegrationServer.java
@@ -10,6 +10,8 @@ import java.io.IOException;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
 import com.google.gson.JsonSyntaxException;
+import group8.Utilities;
+import jdk.jshell.execution.Util;
 import org.eclipse.jetty.server.Server;
 import org.eclipse.jetty.server.Request;
 import org.eclipse.jetty.server.handler.AbstractHandler;
@@ -29,8 +31,8 @@ public class ContinuousIntegrationServer extends AbstractHandler {
         response.setStatus(HttpServletResponse.SC_OK);
         baseRequest.setHandled(true);
 
-        String requestBody = getRequestBodyAsString(request);
-        JsonObject requestJson = deserializeRequest(requestBody);
+        String requestBody = Utilities.getRequestBodyAsString(request);
+        JsonObject requestJson = Utilities.deserializeRequest(requestBody);
 
         System.out.println(requestJson);
         System.out.println(requestJson.get("ref"));
@@ -52,37 +54,4 @@ public class ContinuousIntegrationServer extends AbstractHandler {
         server.start();
         server.join();
     }
-
-    /**
-     * Reads the entire body of an incoming request and returns the result as a string.
-     * @param request - The incoming request
-     * @return the entire request body as a string.
-     * @throws IOException - if the reader fails to read the request body, an IOException will be thrown.
-     */
-    public static String getRequestBodyAsString(HttpServletRequest request) throws IOException {
-        try {
-            StringBuilder requestBuffer = new StringBuilder();
-            BufferedReader requestReader = request.getReader();
-            String nextLine;
-            while ((nextLine = requestReader.readLine()) != null) {
-                requestBuffer.append(nextLine);
-                requestBuffer.append(System.lineSeparator());
-            }
-            return requestBuffer.toString();
-        } catch (IOException err) {
-            return "";
-        }
-    }
-
-    /**
-     * deserializes the incoming request body into a JSON object.
-     * @param requestBody - the request body as a string
-     * @return a JsonObject of the incoming request body
-     * @throws JsonSyntaxException if string passed is invalid as a json object, a JsonSyntaxException is thrown.
-     */
-    public static JsonObject deserializeRequest(String requestBody) throws JsonSyntaxException {
-        JsonObject j = new JsonParser().parse(requestBody).getAsJsonObject();
-        return j;
-    }
-
 }
